Fetcher exceptions are now under a dedicated exception module
